--- Sources/SwiftGodotTestability/EmbeddedXCTest/README.md
@@ -1,3 +1,0 @@
-This code is copied with permission from https://github.com/elegantchaos/EmbeddedXCTest.
-
-Check back there for updates.
